update view holder UI

+++ androidApp/app/src/main/java/com/example/seeStore/controller/CartController.java
@@ -32,9 +32,9 @@ public class CartController {
         return orderItemList;
     }
 
-    public double getSubTotal() {
+    public long getSubTotal() {
         try {
-            double result = 0f;
+            long result = 0l;
 
             for (OrderItem item : orderItemList) {
                 //            result += item.getCost();
@@ -42,7 +42,7 @@ public class CartController {
             return result;
         } catch (Exception e) {
             System.out.println(e.getMessage());
-            return -1f;
+            return -1l;
         }
     }
